feat: Send size information to the client along with data
Previously, we were sending raw data to the client. Now, we send an
integer representing the number of bytes first, then we send the data.

+++ src/session.h
@@ -85,8 +85,8 @@ typedef struct Session {
 
 typedef struct Buffer {
   char *buffer;
-  int n;
-  int capacity;
+  uint64_t n;
+  uint64_t capacity;
 } Buffer;
 
 typedef struct Neotmux {
